+++ source/extensions/isaacsim.asset.exporter.urdf/isaacsim/asset/exporter/urdf/extension.py
@@ -8,102 +8,90 @@
 #
 
 
-import asyncio
 import gc
-import weakref
+from typing import List
 
 import omni.ext
 import omni.ui as ui
 import omni.usd
-from isaacsim.gui.components.menu import make_menu_item_description
+from isaacsim.gui.components.menu import MenuItemDescription
 from omni.kit.menu.utils import add_menu_items, remove_menu_items
-from omni.usd import StageEventType
+from omni.kit.window.file_exporter import ExportOptionsDelegate, get_file_exporter
 
-EXTENSION_TITLE = "USD to URDF Exporter"
-from .ui_builder import UIBuilder
+from .exporter import UrdfExporter
+
+EXTENSION_TITLE = "URDF Exporter"
 
 
 class Extension(omni.ext.IExt):
-    def on_startup(self, ext_id: str):
-        """Initialize extension and UI elements"""
+    def on_startup(self):
+        self._export_options = None
 
-        # Events
-        self._usd_context = omni.usd.get_context()
+        # Menu Setup
+        self._menu_items = [MenuItemDescription(name=EXTENSION_TITLE, onclick_fn=self._show_dialog)]
+        add_menu_items(self._menu_items, "File")
 
-        # Build Window
-        self._window = ui.Window(
-            title=EXTENSION_TITLE, width=600, height=500, visible=False, dockPreference=ui.DockPreference.LEFT_BOTTOM
+    def _show_dialog(self):
+        # File Exporter Dialog setup
+        file_exporter = get_file_exporter()
+        if not file_exporter:
+            return
+
+        self._export_options = UrdfExporterDelegate()
+
+        file_exporter.show_window(
+            title="Export As ...",
+            file_extension_types=[(".urdf", "URDF format")],
+            export_handler=self._export_options.export,
         )
-        self._window.set_visibility_changed_fn(self._on_window)
 
-        # UI
-        self._models = {}
-        self._ext_id = ext_id
-        self._menu_items = [
-            make_menu_item_description(ext_id, EXTENSION_TITLE, lambda a=weakref.proxy(self): a._menu_callback())
-        ]
+        # UrdfExporter specific options inside the file exporter dialog
+        file_exporter.add_export_options_frame("Export Options", self._export_options)
 
-        add_menu_items(self._menu_items, EXTENSION_TITLE)
-
-        # Filled in with User Functions
-        self.ui_builder = UIBuilder()
-
-        # Events
-        self._usd_context = omni.usd.get_context()
-        self._stage_event_sub = None
+    def _hide_dialog(self):
+        file_exporter = get_file_exporter()
+        if file_exporter:
+            file_exporter.hide_window()
 
     def on_shutdown(self):
-        self._models = {}
-        remove_menu_items(self._menu_items, EXTENSION_TITLE)
-        if self._window:
-            self._window = None
-        self.ui_builder.cleanup()
+        # Cleanup Delegate
+        if self._export_options:
+            self._export_options.cleanup()
+        # Cleanup Dialog
+        self._hide_dialog()
+        # Cleanup Menu
+        remove_menu_items(self._menu_items, "File")
+        # Cleanup Garbage
         gc.collect()
 
-    def _on_window(self, visible):
-        if self._window.visible:
-            # Subscribe to Stage Events
-            self._usd_context = omni.usd.get_context()
-            events = self._usd_context.get_stage_event_stream()
-            self._stage_event_sub = events.create_subscription_to_pop(self._on_stage_event)
 
-            self._build_ui()
+class UrdfExporterDelegate(ExportOptionsDelegate):
+    def __init__(self):
+        # Initialize the delegate
+        super().__init__(
+            build_fn=self._build_ui_impl,
+            destroy_fn=self._destroy_impl,
+        )
+        self._widget = None
+        self._exporter = UrdfExporter()
+
+    def _build_ui_impl(self):
+        self._widget = ui.Frame()
+        with self._widget:
+            self._exporter.build_exporter_options()
+
+    def export(self, filename: str, dirname: str, extension: str = "", selections: List[str] = []):
+        result = self._exporter._on_export_button_clicked_fn(dirname, filename)
+        if result:
+            print(f"Export to URDF successful")
         else:
-            self._usd_context = None
-            self._stage_event_sub = None
-            self.ui_builder.cleanup()
+            print(f"Error: Failed to export to URDF")
 
-    def _build_ui(self):
-        with self._window.frame:
-            with ui.VStack(spacing=5, height=0):
-                self._build_extension_ui()
+    def _destroy_impl(self):
+        if self._widget:
+            self._widget.destroy()
+        self._widget = None
 
-        async def dock_window():
-            await omni.kit.app.get_app().next_update_async()
-
-            def dock(space, name, location, pos=0.5):
-                window = omni.ui.Workspace.get_window(name)
-                if window and space:
-                    window.dock_in(space, location, pos)
-                return window
-
-            tgt = ui.Workspace.get_window("Viewport")
-            dock(tgt, EXTENSION_TITLE, omni.ui.DockPosition.LEFT, 0.33)
-            await omni.kit.app.get_app().next_update_async()
-
-        self._task = asyncio.ensure_future(dock_window())
-
-    def _menu_callback(self):
-        self._window.visible = not self._window.visible
-        self.ui_builder.on_menu_callback()
-
-    def _on_stage_event(self, event):
-        if event.type == int(StageEventType.OPENED) or event.type == int(StageEventType.CLOSED):
-            # stage was opened or closed, cleanup
-            self.ui_builder.cleanup()
-
-        self.ui_builder.on_stage_event(event)
-
-    def _build_extension_ui(self):
-        # Call user function for building UI
-        self.ui_builder.build_ui()
+    def cleanup(self):
+        self._exporter.cleanup()
+        self._destroy_impl()