+++ source/extensions/isaacsim.asset.exporter.urdf/docs/CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+
+## [1.2.0] - 2025-04-02
+### Changed
+- Moved Urdf Exporter into File menu next to the other Exporter
+- Options are integrated into the File Exporter dialog
+- Removed option to export from USD file without opening it on Stage
+- usd to urdf library upgraded to python 3.11
+
 ## [1.1.5] - 2025-03-26
 ### Changed
 - Cleanup and standardize extension.toml, update code formatting for all code